requirements.txt
Django==4.2.25
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.11
python-decouple==3.8
gunicorn==21.2.0
channels>=4.0.0
channels-redis>=4.1.0
